renamed screens to pages
