+++ src/layout/Footer.js
@@ -2,11 +2,12 @@ import React from 'react'
 import { Container, Row, Col } from 'react-bootstrap'
 
 const Footer = () => {
+  const year = new Date().getFullYear()
   return (
     <footer>
       <Container>
         <Row>
-          <Col className='text-center py-3'>Copyright &copy; The Shop</Col>
+          <Col className='text-center py-3'>Copyright &copy; {year} The Shop</Col>
         </Row>
       </Container>
     </footer>
