+++ src/layout/Layout.js
@@ -3,12 +3,12 @@ import { Container } from 'react-bootstrap'
 
 import Meta from './Meta'
 import Header from './Header'
-import Footer from '../components/Footer'
+import Footer from './Footer'
 
 const Layout = ({ children, title, description, pageName }) => {
   return (
     <Router>
-      <Meta title={title} pageName={pageName} />
+      <Meta title={title} pageName={pageName} description={description} />
       <Header />
       <Container>
         <main className='py-3'>{children}</main>
