+++ src/pages/HomeScreen.js
@@ -7,7 +7,6 @@ import Message from '../components/Message'
 import Loader from '../components/Loader'
 import Paginate from '../components/Paginate'
 import ProductCarousel from '../components/ProductCarousel'
-import Meta from '../components/Meta'
 import { listProducts } from '../actions/productActions'
 
 const HomeScreen = ({ match }) => {
@@ -26,7 +25,6 @@ const HomeScreen = ({ match }) => {
 
   return (
     <>
-      <Meta />
       {!keyword ? (
         <ProductCarousel />
       ) : (
