+++ src/pages/ProductScreen.js
@@ -5,7 +5,6 @@ import { Row, Col, Image, ListGroup, Card, Button, Form } from 'react-bootstrap'
 import Rating from '../components/Rating'
 import Message from '../components/Message'
 import Loader from '../components/Loader'
-import Meta from '../components/Meta'
 import {
   listProductDetails,
   createProductReview,
@@ -69,7 +68,6 @@ const ProductScreen = ({ history, match }) => {
         <Message variant='danger'>{error}</Message>
       ) : (
         <>
-          <Meta title={product.name} />
           <Row>
             <Col md={6}>
               <Image src={product.image} alt={product.name} fluid />
@@ -223,3 +221,7 @@ const ProductScreen = ({ history, match }) => {
 }
 
 export default ProductScreen
+
+
+{/* <Meta title={product.name} /> */}
+//
